requirements.txt
aryn_sdk==0.2.10
camelot_py==1.0.0
contextily==1.6.2
docxtpl==0.20.0
fitz==0.0.1.dev2
geopandas==1.0.1
matplotlib==3.10.3
numpy==2.3.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
option==2.1.0
pandas==2.3.1
pydantic==2.11.7
pydantic_ai==0.4.9
python_docx==1.1.2
rapidfuzz==3.13.0
Requests==2.32.4
streamlit==1.45.1
streamlit_extras==0.7.5
